move migration utils to a separate file
