+++ python/ray/tune/impl/config.py
@@ -3,8 +3,11 @@ from dataclasses import dataclass
 from ray.air.config import CheckpointConfig as _CheckpointConfig
 from ray.air.config import FailureConfig as _FailureConfig
 from ray.air.config import RunConfig as _RunConfig
-from ray.train.constants import _v2_migration_warnings_enabled
 from ray.train.utils import _copy_doc, _log_deprecation_warning
+from ray.train.v2._internal.migration_utils import (
+    _v2_migration_warnings_enabled,
+    V2_MIGRATION_GUIDE_MESSAGE,
+)
 
 # NOTE: This is just a pass-through wrapper around `ray.train.RunConfig`
 # in order to detect whether the import module was correct (e.g. `ray.tune.RunConfig`).
@@ -36,6 +39,7 @@ class RunConfig(_RunConfig):
                 _log_deprecation_warning(
                     "The `CheckpointConfig` class should be imported from `ray.tune` "
                     "when passing it to the Tuner. Please update your imports."
+                    f"{V2_MIGRATION_GUIDE_MESSAGE}"
                 )
 
         if not isinstance(self.failure_config, FailureConfig):
@@ -43,4 +47,5 @@ class RunConfig(_RunConfig):
                 _log_deprecation_warning(
                     "The `FailureConfig` class should be imported from `ray.tune` "
                     "when passing it to the Tuner. Please update your imports."
+                    f"{V2_MIGRATION_GUIDE_MESSAGE}"
                 )
